+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This document describes notable changes to this project, which adheres
 to [Semantic Versioning](http://semver.org/).
 
 
+
+## Unreleased
+
+### Added
+
+- The `Text` API:
+    - `chance.syllable()`
+
+
 ## [0.2.0] - 2015-08-18
 
 ### Added